requirements.txt
# Core AI and HTTP libraries
anthropic>=0.35.0
python-dotenv>=1.0.0
requests>=2.31.0

# MCP (Model Context Protocol) server integration
mcp>=1.0.0

# Web scraping and content extraction
beautifulsoup4>=4.12.0
firecrawl-py>=0.0.16

# Google Drive integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Token counting and text processing
tiktoken>=0.5.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0 